fix: route Start New Session through the live-session dispatcher instead of a 404 route
StartNew pushed `/practice/live-session-${bundleId}?sessionData=`, which matches no
route (404) and dropped the practice-mode field — a paying Reader could not open a
text chat. It now resolves the selection to concrete phrase ids from the chosen
bundle and routes through the /practice/live-session dispatcher with `mode`, like
the bundle detail page. Extracted the shared pickPhraseIds helper so the two entry
points can't drift. [B2]
